Update Printer model for culling fix
Remove unnecessary face culling that caused visual issues on the block. This fixes an Issue where placing a block against the west face of the block would cause part of the model to disappear. This change also keeps other misc parts from culling when respective side is blocked, despite being able to still see the side.