Add Pull Request Build Action
This commit adds PR.yml which is a GitHub action that will try and build the PR being submitted to make sure it can be built and also to help people beta-testing PRs and hopefully saving everyone time! The built jar files are then uploaded as artifacts of the action in a easy to download .zip file!